+++ src/main/java/party/lemons/biomemakeover/gui/AltarScreen.java
@@ -8,7 +8,6 @@ import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.client.util.math.Vector3f;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.item.ItemStack;
-import net.minecraft.screen.EnchantmentScreenHandler;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.MathHelper;
@@ -52,16 +51,19 @@ public class AltarScreen extends HandledScreen<AltarScreenHandler>
 		drawTexture(matrices, x, y, 0, 0, backgroundWidth, backgroundHeight);
 
 		int progress = this.handler.getProgress();
-		float progressPerc = (float)progress / AltarBlockEntity.MAX_TIME;
-		if (progress > 0) {
-			int n = (int)(28.0F * (1.0F - (float)progress / (float)AltarBlockEntity.MAX_TIME));
-			if (n > 0) {
-				int perc = (int)(progressPerc * 29F);
+		float progressPerc = (float) progress / AltarBlockEntity.MAX_TIME;
+		if(progress > 0)
+		{
+			int n = (int) (28.0F * (1.0F - (float) progress / (float) AltarBlockEntity.MAX_TIME));
+			if(n > 0)
+			{
+				int perc = (int) (progressPerc * 29F);
 				this.drawTexture(matrices, x + 99, y + 55 - perc, 189, 29 - perc, 9, perc);
 			}
 
 			n = GYLPH_PROGRESS[progress / 2 % 13];
-			if (n > 0) {
+			if(n > 0)
+			{
 				this.drawTexture(matrices, x + 68, y + 16 + 53 - n, 177, 53 - n, 12, n);
 			}
 		}
@@ -71,7 +73,7 @@ public class AltarScreen extends HandledScreen<AltarScreenHandler>
 		RenderSystem.matrixMode(5889);
 		RenderSystem.pushMatrix();
 		RenderSystem.loadIdentity();
-		int k = (int)this.client.getWindow().getScaleFactor();
+		int k = (int) this.client.getWindow().getScaleFactor();
 		RenderSystem.viewport((this.width - 320) / 2 * k, (this.height - 240) / 2 * k, 320 * k, 240 * k);
 		RenderSystem.translatef(-0.34F, 0.23F, 0.0F);
 		RenderSystem.multMatrix(Matrix4f.viewboxMatrix(90.0D, 1.3333334F, 9.0F, 80.0F));
@@ -86,27 +88,31 @@ public class AltarScreen extends HandledScreen<AltarScreenHandler>
 		matrices.multiply(Vector3f.POSITIVE_Z.getDegreesQuaternion(180.0F));
 		matrices.multiply(Vector3f.POSITIVE_X.getDegreesQuaternion(20.0F));
 		float g = MathHelper.lerp(delta, this.pageTurningSpeed, this.nextPageTurningSpeed);
-		matrices.translate((1.0F - g) * 0.2F, (1.0F - g) * 0.1F, (double)((1.0F - g) * 0.25F));
+		matrices.translate((1.0F - g) * 0.2F, (1.0F - g) * 0.1F, (1.0F - g) * 0.25F);
 		float h = -(1.0F - g) * 90.0F - 90.0F;
 		matrices.multiply(Vector3f.POSITIVE_Y.getDegreesQuaternion(h));
 		matrices.multiply(Vector3f.POSITIVE_X.getDegreesQuaternion(180.0F));
 		float l = MathHelper.lerp(delta, this.pageAngle, this.nextPageAngle) + 0.25F;
 		float m = MathHelper.lerp(delta, this.pageAngle, this.nextPageAngle) + 0.75F;
-		l = (l - (float)MathHelper.fastFloor(l)) * 1.6F - 0.3F;
-		m = (m - (float)MathHelper.fastFloor(m)) * 1.6F - 0.3F;
-		if (l < 0.0F) {
+		l = (l - (float) MathHelper.fastFloor(l)) * 1.6F - 0.3F;
+		m = (m - (float) MathHelper.fastFloor(m)) * 1.6F - 0.3F;
+		if(l < 0.0F)
+		{
 			l = 0.0F;
 		}
 
-		if (m < 0.0F) {
+		if(m < 0.0F)
+		{
 			m = 0.0F;
 		}
 
-		if (l > 1.0F) {
+		if(l > 1.0F)
+		{
 			l = 1.0F;
 		}
 
-		if (m > 1.0F) {
+		if(m > 1.0F)
+		{
 			m = 1.0F;
 		}
 
@@ -126,14 +132,16 @@ public class AltarScreen extends HandledScreen<AltarScreenHandler>
 	}
 
 	@Override
-	public void render(MatrixStack matrices, int mouseX, int mouseY, float delta) {
+	public void render(MatrixStack matrices, int mouseX, int mouseY, float delta)
+	{
 		renderBackground(matrices);
 		super.render(matrices, mouseX, mouseY, delta);
 		drawMouseoverTooltip(matrices, mouseX, mouseY);
 	}
 
 	@Override
-	protected void init() {
+	protected void init()
+	{
 		super.init();
 		//titleX = (backgroundWidth - textRenderer.getWidth(title)) / 2;
 	}
@@ -145,14 +153,17 @@ public class AltarScreen extends HandledScreen<AltarScreenHandler>
 		doTick();
 	}
 
-	public void doTick() {
+	public void doTick()
+	{
 		ItemStack itemStack = this.handler.getSlot(0).getStack();
-		if (!ItemStack.areEqual(itemStack, this.stack)) {
+		if(!ItemStack.areEqual(itemStack, this.stack))
+		{
 			this.stack = itemStack;
 
-			do {
-				this.approximatePageAngle += (float)(this.random.nextInt(4) - this.random.nextInt(4));
-			} while(this.nextPageAngle <= this.approximatePageAngle + 1.0F && this.nextPageAngle >= this.approximatePageAngle - 1.0F);
+			do
+			{
+				this.approximatePageAngle += (float) (this.random.nextInt(4) - this.random.nextInt(4));
+			}while(this.nextPageAngle <= this.approximatePageAngle + 1.0F && this.nextPageAngle >= this.approximatePageAngle - 1.0F);
 		}
 
 		++this.ticks;
@@ -160,9 +171,11 @@ public class AltarScreen extends HandledScreen<AltarScreenHandler>
 		this.pageTurningSpeed = this.nextPageTurningSpeed;
 		boolean bl = handler.getProgress() > 0;
 
-		if (bl) {
+		if(bl)
+		{
 			this.nextPageTurningSpeed += 0.2F;
-		} else {
+		}else
+		{
 			this.nextPageTurningSpeed -= 0.2F;
 		}
